Move the processors and components into their respetive directories
