+++ packages/contact-form-7/src/components/Input.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FormIdContext from "./FormIdContext";
+import FormIdContext from "./../context/FormIdContext";
 import { connect } from "frontity";
 
 const Input = ({ state, actions, name }) => {
@@ -10,7 +10,7 @@ const Input = ({ state, actions, name }) => {
 
 		console.warn( 'name', id, name, event.target.value );
 
-		state.cf7.forms[id][event.target.name] = event.target.value;
+		// state.cf7.forms[id][event.target.name] = event.target.value;
 		actions.cf7.changeInputValue( id, name, event.target.value );
 	};
