+++ packages/contact-form-7/src/processors/cf7Spans.js
@@ -1,4 +1,5 @@
 import { connect } from "frontity";
+import Span from './../components/Span';
 
 const cf7Spans = {
 	name: "cf7Spans",
